Extracted logic from component
