+++ src/app/ferments/page.js
@@ -1,10 +1,27 @@
+'use client'
 import React from 'react';
+import FermentCard from '../components/FermentCard';
+import { useEffect, useState } from 'react'
 
 
 const Page = () => {
+  const [data, setData] = useState([])
+
+  useEffect(() => {
+    const getFerments = async () => {
+      const res = await fetch('http://127.0.0.1:3002/ferment_recipes')
+      const json = await res.json()
+      setData(json)
+    }
+
+    getFerments()
+  }, [])
+
   return (
     <div className="flex flex-col items-center justify-center min-h-screen py-2">
-      hola soy la ruta de los fermentos
+      {data.map((ferment) => (
+        <FermentCard key={ferment.id} {...ferment} />
+      ))}
     </div>
   );
 }
